Moved getModifiedTableOffset to syncHandler file
