+++ ts/Data/Modifiers/RangeModifier.ts
@@ -220,46 +220,6 @@ class RangeModifier extends DataModifier {
     }
 
 
-    /**
-     * Utility function that returns the first row index
-     * if the table has been modified by a range modifier
-     *
-     * @param {DataTable} table
-     * The table to get the offset from.
-     *
-     * @return {number}
-     * The row offset of the modified table.
-     */
-    public getModifiedTableOffset(
-        table: DataTable
-    ): number {
-        const { ranges } = this.options;
-
-        if (ranges) {
-            const minRange = ranges.reduce(
-                (minRange, currentRange): RangeModifierRangeOptions => {
-                    if (currentRange.minValue > minRange.minValue) {
-                        minRange = currentRange;
-                    }
-                    return minRange;
-
-                }, ranges[0]
-            );
-
-            const tableRowIndex = table.getRowIndexBy(
-                minRange.column,
-                minRange.minValue
-            );
-
-            if (tableRowIndex) {
-                return tableRowIndex;
-            }
-        }
-
-        return 0;
-    }
-
-
 }
